third_party/blink/web_tests/virtual/container-queries/wpt_internal/css/css-conditional/container-queries/inline-size-query-block-size-containment-expected.txt

This is a testharness.js-based test.
FAIL Axis mismatch between query and size containment assert_equals: expected 50 but got 400
Harness: the test ran to completion.
